+++ cypress.config.js
@@ -1,38 +1,32 @@
-const { defineConfig }     = require('cypress')
-const { loadEnv }          = require('vite')
-const { default:cucumber } = require('cypress-cucumber-preprocessor')
+const path             = require('path')
+const { defineConfig } = require('cypress')
+const { loadEnv }      = require('vite')
+
+const createBundler                     = require('@bahmutov/cypress-esbuild-preprocessor')
+const { addCucumberPreprocessorPlugin } = require('@badeball/cypress-cucumber-preprocessor')
+const { createEsbuildPlugin }           = require('@badeball/cypress-cucumber-preprocessor/esbuild')
+
+const visualTestingTasks = require('heimdall-visual-test/src/cypress/taskHandlers')
 
 const envTestFile = loadEnv('test', process.cwd(), '')
 const port        = envTestFile.PORT || '3000'
 const baseUrl     = `http://localhost:${port}`
 
 module.exports = defineConfig({
-  viewportWidth:              1336,
-  viewportHeight:             768,
-  eyesIsDisabled:             false,
-  eyesFailCypressOnDiff:      true,
-  eyesDisableBrowserFetching: false,
-  eyesTestConcurrency:        5,
-  eyesIsGlobalHooksSupported: false,
-  eyesPort:                   64361,
-  appliConfFile: {
-    batch: {
-      id: '0d583209-12ad-4192-93a3-deb1df9b55f9',
-    },
-  },
+  viewportWidth:  1336,
+  viewportHeight: 768,
   e2e: {
     baseUrl,
     specPattern: 'cypress/e2e/**/*.{feature,features}',
-    setupNodeEvents(on, config) {
-      on('file:preprocessor', cucumber())
+    async setupNodeEvents(on, config) {
+      await addCucumberPreprocessorPlugin(on, config)
 
+      on('file:preprocessor', createBundler({plugins: [createEsbuildPlugin(config)]}))
+      on('task', {...visualTestingTasks(config)})
       config.env = {...config.env, ...envTestFile}
 
-      const mockModule = {
-        exports: (on, config) => { }
-      }
-      require('@applitools/eyes-cypress')(mockModule)
-      return mockModule.exports(on, config)
+      return config
     },
+    visualTestFolder: path.join(__dirname, 'cypress/visual-test-screenshots'),
   },
 })